Fixed bugs caused by onpause-onresume
* Unbinding moved to a specific method pauseCurActivity
* This method is called when backbutton is pressed and when new activity
is getting spawned. It's important to put them here and not in the
onStop() method because that gets called after new activity is spawned
and messes with the unbinding thing.
* Rebinding moved on OnRestart method, which gets called when activity
resumes from backstack on pressing back button.
* added trycatch in isPng() method to prevent crashes from
illegalstateexception. Not sure if this is fixing the issue, with
isPng() called in wrong state on UI thread, but it seems to.